scripts/requirements.txt
krcg
pyyaml
yamlfix
yamllint
unidecode
Faker
